+++ core/src/com/mygdx/game/Views/ViewManager.java
@@ -1,6 +1,8 @@
 package com.mygdx.game.Views;
 
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
 import java.util.Stack;
@@ -8,9 +10,13 @@ import java.util.Stack;
 public class ViewManager {
 
     private Stack<View> views;
+    private Music music;
 
     public ViewManager() {
         views = new Stack<>();
+        music = Gdx.audio.newMusic(Gdx.files.internal("music.mp3"));
+        music.setLooping(true);
+        music.play();
     }
 
     public void push(View state) {
@@ -38,5 +44,14 @@ public class ViewManager {
 
     public void render(SpriteBatch sb) {
         views.peek().render(sb);
+    }
+
+    public void setMusic(boolean play) {
+        if (play){
+            music.play();
+        }
+        else{
+            music.pause();
+        }
     }
 }
